extract depinject config
