+++ core/appmodule/README.md
@@ -4,5 +4,4 @@
 
 This package defines what is needed for an module to be used in the Cosmos SDK. 
 
-
-If you are looking at integrating Dependency injection into your module please see [depinject appmodule documentation](../../depinject/appmodule/README.md)
+If you are looking at integrating Dependency injection into your module please see [depinject appconfig documentation](../../depinject/appconfig/README.md)
