+++ orm/model/ormdb/module_test.go
@@ -13,14 +13,12 @@ import (
 	"gotest.tools/v3/assert"
 	"gotest.tools/v3/golden"
 
-	appv1alpha1 "cosmossdk.io/api/cosmos/app/v1alpha1"
 	ormmodulev1alpha1 "cosmossdk.io/api/cosmos/orm/module/v1alpha1"
 	ormv1alpha1 "cosmossdk.io/api/cosmos/orm/v1alpha1"
 	"cosmossdk.io/core/genesis"
 	"cosmossdk.io/core/store"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/depinject/appconfig"
-	am "cosmossdk.io/depinject/appmodule"
 	_ "cosmossdk.io/orm" // required for ORM module registration
 	"cosmossdk.io/orm/internal/testkv"
 	"cosmossdk.io/orm/internal/testpb"
@@ -400,8 +398,8 @@ func ProvideTestRuntime() store.KVStoreService {
 }
 
 func TestAppConfigModule(t *testing.T) {
-	appCfg := appconfig.Compose(&appv1alpha1.Config{
-		Modules: []*appv1alpha1.ModuleConfig{
+	appCfg := appconfig.Compose(&appconfig.AppConfig{
+		Modules: []*appconfig.ModuleConfig{
 			{Name: "bank", Config: appconfig.WrapAny(&testpb.Module{})},
 			{Name: "orm", Config: appconfig.WrapAny(&ormmodulev1alpha1.Module{})},
 		},
