+++ x/accounts/depinject.go
@@ -1,0 +1,8 @@
+package accounts
+
+import "cosmossdk.io/depinject"
+
+var _ depinject.OnePerModuleType = AppModule{}
+
+// IsOnePerModuleType implements the depinject.OnePerModuleType interface.
+func (am AppModule) IsOnePerModuleType() {}
