+++ x/circuit/module.go
@@ -8,21 +8,13 @@ import (
 
 	gwruntime "github.com/grpc-ecosystem/grpc-gateway/runtime"
 
-	modulev1 "cosmossdk.io/api/cosmos/circuit/module/v1"
-	"cosmossdk.io/core/address"
 	"cosmossdk.io/core/appmodule"
-	"cosmossdk.io/core/store"
-	"cosmossdk.io/depinject"
-	am "cosmossdk.io/depinject/appmodule"
-	authtypes "cosmossdk.io/x/auth/types"
 	"cosmossdk.io/x/circuit/keeper"
 	"cosmossdk.io/x/circuit/types"
 
-	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-	"github.com/cosmos/cosmos-sdk/runtime"
 	"github.com/cosmos/cosmos-sdk/telemetry"
 	"github.com/cosmos/cosmos-sdk/types/module"
 )
@@ -85,9 +77,6 @@ type AppModule struct {
 	keeper keeper.Keeper
 }
 
-// IsOnePerModuleType implements the depinject.OnePerModuleType interface.
-func (am AppModule) IsOnePerModuleType() {}
-
 // IsAppModule implements the appmodule.AppModule interface.
 func (am AppModule) IsAppModule() {}
 
@@ -124,51 +113,4 @@ func (am AppModule) InitGenesis(ctx context.Context, cdc codec.JSONCodec, data j
 func (am AppModule) ExportGenesis(ctx context.Context, cdc codec.JSONCodec) json.RawMessage {
 	gs := am.keeper.ExportGenesis(ctx)
 	return cdc.MustMarshalJSON(gs)
-}
-
-func init() {
-	am.Register(
-		&modulev1.Module{},
-		am.Provide(ProvideModule),
-	)
-}
-
-type ModuleInputs struct {
-	depinject.In
-
-	Config       *modulev1.Module
-	Cdc          codec.Codec
-	StoreService store.KVStoreService
-
-	AddressCodec address.Codec
-}
-
-type ModuleOutputs struct {
-	depinject.Out
-
-	CircuitKeeper  keeper.Keeper
-	Module         appmodule.AppModule
-	BaseappOptions runtime.BaseAppOption
-}
-
-func ProvideModule(in ModuleInputs) ModuleOutputs {
-	// default to governance authority if not provided
-	authority := authtypes.NewModuleAddress("gov")
-	if in.Config.Authority != "" {
-		authority = authtypes.NewModuleAddressOrBech32Address(in.Config.Authority)
-	}
-
-	circuitkeeper := keeper.NewKeeper(
-		in.Cdc,
-		in.StoreService,
-		authority.String(),
-		in.AddressCodec,
-	)
-	m := NewAppModule(in.Cdc, circuitkeeper)
-
-	baseappOpt := func(app *baseapp.BaseApp) {
-		app.SetCircuitBreaker(&circuitkeeper)
-	}
-
-	return ModuleOutputs{CircuitKeeper: circuitkeeper, Module: m, BaseappOptions: baseappOpt}
 }
