+++ x/params/module.go
@@ -5,14 +5,8 @@ import (
 
 	gwruntime "github.com/grpc-ecosystem/grpc-gateway/runtime"
 
-	modulev1 "cosmossdk.io/api/cosmos/params/module/v1"
 	"cosmossdk.io/core/appmodule"
-	"cosmossdk.io/depinject"
-	am "cosmossdk.io/depinject/appmodule"
-	store "cosmossdk.io/store/types"
-	govv1beta1 "cosmossdk.io/x/gov/types/v1beta1"
 	"cosmossdk.io/x/params/keeper"
-	"cosmossdk.io/x/params/types"
 	"cosmossdk.io/x/params/types/proposal"
 
 	"github.com/cosmos/cosmos-sdk/client"
@@ -26,7 +20,6 @@ var (
 	_ module.AppModuleBasic      = AppModule{}
 	_ module.AppModuleSimulation = AppModule{}
 	_ module.HasServices         = AppModule{}
-	_ depinject.OnePerModuleType = AppModule{}
 
 	_ appmodule.AppModule = AppModule{}
 )
@@ -73,9 +66,6 @@ func NewAppModule(k keeper.Keeper) AppModule {
 	}
 }
 
-// IsOnePerModuleType implements the depinject.OnePerModuleType interface.
-func (am AppModule) IsOnePerModuleType() {}
-
 // IsAppModule implements the appmodule.AppModule interface.
 func (am AppModule) IsAppModule() {}
 
@@ -98,58 +88,3 @@ func (am AppModule) WeightedOperations(_ module.SimulationState) []simtypes.Weig
 
 // ConsensusVersion implements AppModule/ConsensusVersion.
 func (AppModule) ConsensusVersion() uint64 { return ConsensusVersion }
-
-//
-// App Wiring Setup
-//
-
-func init() {
-	am.Register(&modulev1.Module{},
-		am.Provide(
-			ProvideModule,
-			ProvideSubspace,
-		))
-}
-
-type ModuleInputs struct {
-	depinject.In
-
-	KvStoreKey        *store.KVStoreKey
-	TransientStoreKey *store.TransientStoreKey
-	Cdc               codec.Codec
-	LegacyAmino       *codec.LegacyAmino
-}
-
-type ModuleOutputs struct {
-	depinject.Out
-
-	ParamsKeeper keeper.Keeper
-	Module       appmodule.AppModule
-	GovHandler   govv1beta1.HandlerRoute
-}
-
-func ProvideModule(in ModuleInputs) ModuleOutputs {
-	k := keeper.NewKeeper(in.Cdc, in.LegacyAmino, in.KvStoreKey, in.TransientStoreKey)
-
-	m := NewAppModule(k)
-	govHandler := govv1beta1.HandlerRoute{RouteKey: proposal.RouterKey, Handler: NewParamChangeProposalHandler(k)}
-
-	return ModuleOutputs{ParamsKeeper: k, Module: m, GovHandler: govHandler}
-}
-
-type SubspaceInputs struct {
-	depinject.In
-
-	Key       depinject.ModuleKey
-	Keeper    keeper.Keeper
-	KeyTables map[string]types.KeyTable
-}
-
-func ProvideSubspace(in SubspaceInputs) types.Subspace {
-	moduleName := in.Key.Name()
-	kt, exists := in.KeyTables[moduleName]
-	if !exists {
-		return in.Keeper.Subspace(moduleName)
-	}
-	return in.Keeper.Subspace(moduleName).WithKeyTable(kt)
-}
