+++ x/upgrade/module.go
@@ -6,27 +6,17 @@ import (
 	"fmt"
 
 	gwruntime "github.com/grpc-ecosystem/grpc-gateway/runtime"
-	"github.com/spf13/cast"
 	"github.com/spf13/cobra"
 
-	modulev1 "cosmossdk.io/api/cosmos/upgrade/module/v1"
 	"cosmossdk.io/core/address"
 	"cosmossdk.io/core/appmodule"
-	"cosmossdk.io/core/store"
-	"cosmossdk.io/depinject"
-	am "cosmossdk.io/depinject/appmodule"
-	authtypes "cosmossdk.io/x/auth/types"
 	"cosmossdk.io/x/upgrade/client/cli"
 	"cosmossdk.io/x/upgrade/keeper"
 	"cosmossdk.io/x/upgrade/types"
 
-	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-	"github.com/cosmos/cosmos-sdk/server"
-	servertypes "github.com/cosmos/cosmos-sdk/server/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 )
 
@@ -89,9 +79,6 @@ func NewAppModule(keeper *keeper.Keeper, ac address.Codec) AppModule {
 		keeper:         keeper,
 	}
 }
-
-// IsOnePerModuleType implements the depinject.OnePerModuleType interface.
-func (am AppModule) IsOnePerModuleType() {}
 
 // IsAppModule implements the appmodule.AppModule interface.
 func (am AppModule) IsAppModule() {}
@@ -159,74 +146,4 @@ func (AppModule) ConsensusVersion() uint64 { return ConsensusVersion }
 // CONTRACT: this is called *before* all other modules' BeginBlock functions
 func (am AppModule) PreBlock(ctx context.Context) (appmodule.ResponsePreBlock, error) {
 	return PreBlocker(ctx, am.keeper)
-}
-
-//
-// App Wiring Setup
-//
-
-func init() {
-	am.Register(&modulev1.Module{},
-		am.Provide(ProvideModule),
-		am.Invoke(PopulateVersionMap),
-	)
-}
-
-type ModuleInputs struct {
-	depinject.In
-
-	Config             *modulev1.Module
-	StoreService       store.KVStoreService
-	Cdc                codec.Codec
-	AddressCodec       address.Codec
-	AppVersionModifier baseapp.AppVersionModifier
-
-	AppOpts servertypes.AppOptions `optional:"true"`
-}
-
-type ModuleOutputs struct {
-	depinject.Out
-
-	UpgradeKeeper *keeper.Keeper
-	Module        appmodule.AppModule
-}
-
-func ProvideModule(in ModuleInputs) ModuleOutputs {
-	var (
-		homePath           string
-		skipUpgradeHeights = make(map[int64]bool)
-	)
-
-	if in.AppOpts != nil {
-		for _, h := range cast.ToIntSlice(in.AppOpts.Get(server.FlagUnsafeSkipUpgrades)) {
-			skipUpgradeHeights[int64(h)] = true
-		}
-
-		homePath = cast.ToString(in.AppOpts.Get(flags.FlagHome))
-	}
-
-	// default to governance authority if not provided
-	authority := authtypes.NewModuleAddress(types.GovModuleName)
-	if in.Config.Authority != "" {
-		authority = authtypes.NewModuleAddressOrBech32Address(in.Config.Authority)
-	}
-
-	auth, err := in.AddressCodec.BytesToString(authority)
-	if err != nil {
-		panic(err)
-	}
-
-	// set the governance module account as the authority for conducting upgrades
-	k := keeper.NewKeeper(skipUpgradeHeights, in.StoreService, in.Cdc, homePath, in.AppVersionModifier, auth)
-	m := NewAppModule(k, in.AddressCodec)
-
-	return ModuleOutputs{UpgradeKeeper: k, Module: m}
-}
-
-func PopulateVersionMap(upgradeKeeper *keeper.Keeper, modules map[string]appmodule.AppModule) {
-	if upgradeKeeper == nil {
-		return
-	}
-
-	upgradeKeeper.SetInitVersionMap(module.NewManagerFromMap(modules).GetVersionMap())
 }
